requirements.txt
discord
python-dotenv
pynacl
